chatbot-api/requirements.txt
streamlit
plotly
pandas
flask
gunicorn
boto3
python-dotenv
numpy